pyproject.toml
[project]
name = "macos-mlx-benchmark"
version = "0.1.0"
description = "MLX inference benchmarking tool for Apple Silicon"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "mlx-lm>=0.21",
    "mlx-vlm>=0.1",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "textual>=1.0",
    "zeus[apple]",
]

[project.scripts]
bench = "bench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bench"]
